\name{stdz}
\alias{stdz}
\title{ Utility Function to Standardize (Make to Have Norm One) Basis Functions }
\description{
  The purpose of this function is to numerically stablize the computation.
}
\usage{
stdz(x)
}
\arguments{
  \item{x}{ Matrix of basis functions. Each column of which is a basis. }
}
\keyword{ smooth }
